android adjustments
