+++ android/app/build.gradle
@@ -236,11 +236,12 @@ android {
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "com.facebook.react:react-native:+"  // From node_modules
-    implementation project(':react-native-notifications')
-    implementation 'com.google.firebase:firebase-core:16.0.0'
+    implementation 'com.google.firebase:firebase-core:17.4.0'
     implementation 'com.google.firebase:firebase-messaging:21.1.0'
     implementation 'com.google.firebase:firebase-analytics:17.4.0'
     implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.0.0"
+    implementation project(':react-native-notifications')
+    implementation project(':react-native-push-notification')
     implementation 'org.chromium:v8-android:9.93.+'
 
     implementation ("androidx.appcompat:appcompat:1.3.1") {
