+++ android/build.gradle
@@ -9,7 +9,6 @@ buildscript {
         ndkVersion = "20.1.5948944"
         ext.kotlinVersion = '1.3.41'
         firebaseMessagingVersion = "21.1.0"
-        googlePlayServicesVersion = "17.5.0"
     }
     repositories {
         google()
@@ -28,6 +27,7 @@ buildscript {
 
 allprojects {
     repositories {
+        google()
         mavenLocal()
         maven {
             // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
@@ -54,7 +54,6 @@ allprojects {
         maven { url "$rootDir/../node_modules/detox/Detox-android"}
 
         mavenCentral()
-        google()
         jcenter()
 
         maven { url 'https://www.jitpack.io' }
